refactor: doamin module
