+++ src/domain/repositories/movie.repositoy.ts
@@ -1,4 +1,4 @@
-import { Movie } from '@/domain/mskmovies/movie';
+import { Movie } from '@/domain/model/movie';
 import { UpdateMovieDto } from '@/infra/http/dto';
 
 export abstract class MovieRepository {
