+++ src/infra/http/movies.controller.ts
@@ -15,12 +15,12 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { CacheInterceptor, CacheKey } from '@nestjs/cache-manager';
-import { CreateMovieUseCase } from '@/application/mskmovies/use-case/create-movie';
+import { CreateMovieUseCase } from '@/domain/use-case/create-movie';
 import { CreateMovieDto, UpdateMovieDto } from './dto';
-import { GetMoviesUseCase } from '@/application/mskmovies/use-case/get-movies';
-import { DeleteMovieUseCase } from '@/application/mskmovies/use-case/delete-movie';
-import { GetMovieUseCase } from '@/application/mskmovies/use-case/get-movie';
-import { UpdateMovieUseCase } from '@/application/mskmovies/use-case/update-movie';
+import { GetMoviesUseCase } from '@/domain/use-case/get-movies';
+import { DeleteMovieUseCase } from '@/domain/use-case/delete-movie';
+import { GetMovieUseCase } from '@/domain/use-case/get-movie';
+import { UpdateMovieUseCase } from '@/domain/use-case/update-movie';
 
 @ApiTags('Movies')
 @ApiBearerAuth()
