+++ src/infra/persistence/typeorm/repositories/typeorm-movies.repository.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { MovieEntity } from '../entities/movie.entity';
-import { Movie } from '@/domain/mskmovies/movie';
+import { Movie } from '@/domain/model/movie';
 import { TypeOrmMovieMapper } from '../mapper/typeorm-movie-mapper';
-import { MovieRepository } from '@/application/mskmovies/ports/movie.repositoy';
+import { MovieRepository } from '@/domain/repositories/movie.repositoy';
 
 @Injectable()
 export class TypeOrmMoviesRepository implements MovieRepository {
